pulled out hookup dossier to separate file, changed check parts and some import cleanup
