+++ hera_mc/cm_sysutils.py
@@ -192,7 +192,7 @@ class Handling:
             H = self.H
         else:
             H = cm_hookup.Hookup(self.session)
-        hud = H.get_hookup(hpn_list=[stn], at_date=at_date, exact_match=True, hookup_type=hookup_type)
+        hud = H.get_hookup(hpn=[stn], at_date=at_date, exact_match=True, hookup_type=hookup_type)
         station_info = None
         fully_connected = cm_revisions.get_full_revision(stn, hud)
         fully_connected_keys = set()
@@ -342,7 +342,7 @@ class Handling:
         H = cm_hookup.Hookup(self.session)
         if isinstance(stn, six.string_types):
             stn = [stn]
-        hud = H.get_hookup(hpn_list=stn, at_date=at_date, exact_match=True, hookup_type=hookup_type)
+        hud = H.get_hookup(hpn=stn, at_date=at_date, exact_match=True, hookup_type=hookup_type)
         for k, hu in six.iteritems(hud):
             parts[k] = hu.get_part_in_hookup_from_type(part_type, include_revs=include_revs, include_ports=include_ports)
         return parts
@@ -371,13 +371,12 @@ class Handling:
             Status string.  "OK" or "Not on 'main'"
         """
         import os.path
-        if isinstance(hlist, six.string_types):
-            hlist = [hlist]
-        if hlist[0].lower() == 'default':
+        if isinstance(hlist, six.string_types) and hlist.lower() == 'default':
             hlist = cm_utils.default_station_prefixes
+        hlist = cm_utils.listify(hlist)
         output_file = os.path.expanduser('~/.hera_mc/sys_conn_tmp.html')
         H = cm_hookup.Hookup(self.session)
-        hookup_dict = H.get_hookup(hpn_list=hlist, rev=rev, port_query='all', at_date='now', exact_match=exact_match,
+        hookup_dict = H.get_hookup(hpn=hlist, pol='all', at_date='now', exact_match=exact_match,
                                    force_new_cache=force_new_hookup_dict, force_db=False, hookup_type=None)
         H.show_hookup(hookup_dict=hookup_dict, cols_to_show=hookup_cols, ports=True,
                       revs=True, state='full', filename=output_file, output_format='html')
